requirements.txt
streamlit==1.37.0
streamlit-webrtc==0.47.1

requests
pydub
pyyaml

groq